🧪 15G7ZE task: split human input lifecycle test
